% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_data.R
\name{acs_mode_to_work_county}
\alias{acs_mode_to_work_county}
\title{PSRC Counties and Region ACS Work Mode Share}
\usage{
acs_mode_to_work_county(years)
}
\arguments{
\item{years}{Four digit integer (or vector of integers) for year(s) of analysis}
}
\value{
tibble of mode to work data by county and region by census year
}
\description{
This function pulls and cleans ACS Mode Share Data from Census Table B08006.
Pulls data by year for King, Kitsap, Pierce and Snohomish counties and summarized by PSRC Region
}
\examples{
modes_psrc <- acs_mode_to_work_county(years <- 2021)

}
