% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_data.R
\name{acs_mode_to_work_place}
\alias{acs_mode_to_work_place}
\title{PSRC Cities ACS Work Mode Share}
\usage{
acs_mode_to_work_place(years)
}
\arguments{
\item{years}{Four digit integer (or vector of integers) for year(s) of analysis}
}
\value{
tibble of mode to work data by county and region by census year
}
\description{
This function pulls and cleans ACS Mode Share Data from Census Table B08006.
Pulls data by year for cities in the PSRC Region
}
\examples{
modes_psrc <- acs_mode_to_work_place(years <- 2020)

}
