% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_data.R
\name{blocks_served_by_transit}
\alias{blocks_served_by_transit}
\title{Census Blocks served by Transit}
\usage{
blocks_served_by_transit(
  year,
  service_change = "Spring",
  buffer = 0.5,
  modes = "HCT"
)
}
\arguments{
\item{year}{Year as a four digit integer - available from 2015 to current year}

\item{service_change}{Service Change period - either Fall or Spring - defaults to Spring}

\item{buffer}{Buffer Distance express in miles - defaults to 0.5}

\item{modes}{Either HCT or All for transit serving the blocks - defaults to HCT}
}
\value{
character vector of census block id's served by transit modes
}
\description{
This function pulls GTFS data by Mode and intersect with Blocks to determine which blocks are served by transit.
}
\examples{
\dontrun{
hct_blocks <- blocks_served_by_transit(year=2022)}

}
