% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_data.R
\name{equity_population_near_transit}
\alias{equity_population_near_transit}
\title{Equity Population served by Transit}
\usage{
equity_population_near_transit(
  year,
  service_change = "Spring",
  modes = "HCT",
  latest_census_year = 2021,
  buffer = 0.5
)
}
\arguments{
\item{year}{Year as a four digit integer - available from 2015 to current year}

\item{service_change}{Service Change period - either Fall or Spring - defaults to Spring}

\item{modes}{Either HCT or All for transit serving the blocks - defaults to HCT}

\item{latest_census_year}{Four digit integer for latest ACS 5 yr data - defaults to 2021}

\item{buffer}{Buffer Distance express in miles - defaults to 0.50}
}
\value{
tibble of population and share of people serve by transit by equity focus population
}
\description{
This function pulls GTFS data by Mode, intersects with Blocks and calculates how many people are served by transit.
}
\examples{
\dontrun{
pop_half_mile_hct <- equity_population_near_transit(year=2022)
pop_qtr_mile_hct <- equity_population_near_transit(year=2022, buffer=0.25)}

}
