% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_registrations.R
\name{ev_registrations_by_county}
\alias{ev_registrations_by_county}
\title{Total Vehicle Registrations by Electrification Level and County}
\usage{
ev_registrations_by_county()
}
\value{
tibble of total vehicle registrations by electrification level by month and county
}
\description{
This function pulls and cleans data from the Washington State Registration Data.
Data is pulled from https://data.wa.gov/ via the Socrata API
}
\examples{
\dontrun{
total_vehicle_registrations <- ev_registrations_by_county()}

}
