% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_registrations.R
\name{ev_registrations_tract}
\alias{ev_registrations_tract}
\title{New Vehicle Registrations by Electrification Level and Census Tract}
\usage{
ev_registrations_tract(
  data_file,
  title_type = "Original Title",
  vehicle_type = "New"
)
}
\arguments{
\item{data_file}{path to downloaded data files https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x}

\item{title_type}{Either an "Original Title" or a "Transfer Title" - defaults to "Original Title"}

\item{vehicle_type}{Either for "New" or "Used" vehicles  -defaults to "New"}
}
\value{
tibble of new vehicle registrations by electrification level by month and county
}
\description{
This cleans data from the Washington State Registration Data vehicle registrations on open data portal.
The data is pre-downloaded due to size from: https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x
}
\examples{
\dontrun{
registrations_by_tract <- ev_registrations_tract(data_file="C:/coding/Vehicle_Title_Transactions.csv", 
                                                 title_type=c("Original Title"), 
                                                 vehicle_type=c("New"))}

}
