% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{get_ofm_housing_data}
\alias{get_ofm_housing_data}
\title{Get OFM Housing Data by Jurisdiction}
\usage{
get_ofm_housing_data()
}
\value{
tibble of population by jurisdiction
}
\description{
This function pulls Housing Data by Jurisdiction.
}
\examples{
\dontrun{
ofm_housing_data <- get_ofm_housing_data()}

}
