% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{get_ofm_pop_data}
\alias{get_ofm_pop_data}
\title{Get OFM Intercensal Population Data by Jurisdiction}
\usage{
get_ofm_pop_data()
}
\value{
tibble of population by jurisdiction
}
\description{
This function pulls Intercensal Population Data by Jurisdiction.
}
\examples{

ofm_pop_data <- get_ofm_pop_data()

}
