% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{housing_units_data}
\alias{housing_units_data}
\title{Housing Trends from the Office of Financial Management}
\usage{
housing_units_data(forecast_base_yr = 2018)
}
\arguments{
\item{forecast_base_yr}{Four digit integer for Base Year in Forecast Data - defaults to 2018}
}
\value{
tibble of housing trends for the region and regional geographies by calendar year
}
\description{
This function pulls and cleans data from the Intercensal Population data from OFM.
Forecast data is from the latest VISION 2050 modleing as stored on the file server
}
\examples{

housing_units <- housing_units_data()

}
