% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_data.R
\name{jobs_forecast_data}
\alias{jobs_forecast_data}
\title{Total Observed and Forecast Employment Growth}
\usage{
jobs_forecast_data(forecast_base_yr = 2018)
}
\arguments{
\item{forecast_base_yr}{Four digit integer for Base Year in Forecast Data - defaults to 2018}
}
\value{
tibble of total jobs in the region by calendar year
}
\description{
This function pulls total employment estimates from PSRC's employment database in Elmer and the Macroforecast.
The jobs estimates are already summarized by year in a table in Elmer.
}
\examples{
\dontrun{
forecast_jobs <- jobs_forecast_data()}
 
}
