% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npmrds_data.R
\name{map_npmrds_data}
\alias{map_npmrds_data}
\title{Am, Midday and PM Speed Ratios for mapping}
\usage{
map_npmrds_data(file_path = "X:/DSA/rtp-dashboard/NPMRDS/")
}
\arguments{
\item{file_path}{Absolute file path to pre-processed travel time files from NPMRDS}
}
\value{
tibble am, midday and pm speed ratios by TMC with geometry for mapping
}
\description{
This function cleans data from the NPMRDS output files from INRIX.
Data is pre-processed through a python script on data downloaded from the RTIS site
}
\examples{
\dontrun{
congestion_map_data <- map_npmrds_data()}

}
