% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_registrations.R
\name{new_ev_models}
\alias{new_ev_models}
\title{Electric Vehicle Makers by Vehicle Registration}
\usage{
new_ev_models(data_file = "C:/coding/Vehicle_Title_Transactions.csv")
}
\arguments{
\item{data_file}{path to downloaded data files https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x}
}
\value{
tibble of electric vehicle manufacturers registered in the region by month
}
\description{
This cleans data from the Washington State Registration Data vehicle registrations on open data portal.
The data is pre-downloaded due to size from: https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x
}
\examples{
\dontrun{
ev_manufacturers <- new_ev_models(data_file="C:/coding/Vehicle_Title_Transactions.csv")}

}
