% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_registrations.R
\name{new_ev_registrations}
\alias{new_ev_registrations}
\title{New Vehicle Registrations by Electrification Level}
\usage{
new_ev_registrations(
  data_file,
  title_type = "Original Title",
  vehicle_type = "New"
)
}
\arguments{
\item{data_file}{path to downloaded data files https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x}

\item{title_type}{Either an "Original Title" or a "Transfer Title" - defaults to "Original Title"}

\item{vehicle_type}{Either for "New" or "Used" vehicles  -defaults to "New"}
}
\value{
tibble of new vehicle registrations by electrification level by month and county
}
\description{
This cleans data from the Washington State Registration Data vehicle registrations on open data portal.
The data is pre-downloaded ue to size from: https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x
}
\examples{
\dontrun{
new_vehicle_original_registrations <- new_ev_registrations(data_file="C:/coding/Vehicle_Title_Transactions.csv", 
                                                           title_type=c("Original Title"), 
                                                           vehicle_type=c("New"))}

}
