% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_registrations.R
\name{new_vehicle_registrations}
\alias{new_vehicle_registrations}
\title{New Vehicle Registrations by Electrification Level}
\usage{
new_vehicle_registrations(
  mo = c(1, 4, 7, 10),
  yrs = c(2020, 2021, 2022),
  counties = c("King", "Kitsap", "Pierce", "Snohomish")
)
}
\arguments{
\item{mo}{List of months as integers - defaults 1,4,7 and 10}

\item{yrs}{List of years as 4 digit integers from 2020 onward - defaults to 2020-2022}

\item{counties}{List of County Names - defaults to PSRC counties}
}
\value{
tibble of new vehicle registrations by electrification level by month and county
}
\description{
This function pulls and cleans data from the Washington State Registration Data.
Data is pulled from https://data.wa.gov/ via the Socrata API
}
\examples{

new_veh <- new_vehicle_registrations(mo=c(4), yrs=c(2022))

}
