% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{population_by_efa}
\alias{population_by_efa}
\title{Block Level Estimates of equity Focus Populations}
\usage{
population_by_efa(year, latest_census_year = 2021)
}
\arguments{
\item{year}{Four digit integer for current year of analysis}

\item{latest_census_year}{Four digit integer for latest ACS 5 yr data - defaults to 2021}
}
\value{
tibble of population by equity focus area by census block by calendar year
}
\description{
This function pulls and cleans data from the block level Population data from OFM.
Uses Census Tracts or Block Groups to assign shares of the 6 equity focus areas to blocks
}
\examples{

pop_by_block_efa <- population_by_efa(year=2022)

}
