% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{population_data}
\alias{population_data}
\title{Population and Housing Trends from the Office of Financial Management}
\usage{
population_data(forecast_base_yr = 2018)
}
\arguments{
\item{forecast_base_yr}{Four digit integer for Base Year in Forecast Data - defaults to 2018}
}
\value{
tibble of population and housing trends for the region and regional geographies by calendar year
}
\description{
This function pulls and cleans data from the Intercensal Population data from OFM.
Forecast data is from the latest PSRC Macroeconomic forecast as stored in Elmer
}
\examples{

pop_data<- population_data()

}
