% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{population_near_hct}
\alias{population_near_hct}
\title{Population and Housing Unit Growth Near High Capacity Transit}
\usage{
population_near_hct()
}
\value{
tibble of population and housing growth near High Capacity Transit by calendar year
}
\description{
This function pulls and cleans data from the SAEP Block data from OFM for 2020 onward.
2010 to 2020 SAEP Block Data is stored in Elmer
VISION 2050 Station Area and RGC buffers are stored in ElmerGeo
}
\examples{
\dontrun{
pop_hsg_hct <- population_near_hct()}

 
}
