% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\name{process_commute_data}
\alias{process_commute_data}
\title{Commute Travel Mode and Time Data from ACS & PUMS Data}
\usage{
process_commute_data(data_years = c(2021))
}
\arguments{
\item{data_years}{List of data years. Each year should be the latest in non-overlapping five-year series.}
}
\value{
tibble ACS and PUMS travel mode, time and departure time to work by race/ethnicity, county, and region
}
\description{
This function processes Public Use Microdata Sample (PUMS) ACS etimates from the U.S. Census Bureau.
Data is pulled using PSRC's psrccensus package.
}
\examples{
\dontrun{
commute_data <- process_commute_data(data_years = c(2021))}

}
