% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatal_crash_data.R
\name{process_crash_data_annual}
\alias{process_crash_data_annual}
\title{Traffic Fatalities from the Washington Traffic Safety Commission}
\usage{
process_crash_data_annual(data_years = seq(10, 20, by = 1))
}
\arguments{
\item{data_years}{List of two-digit integers for years of analysis - defaults to 10 to 20}
}
\value{
tibble of fatal collision metrics for the region and counties by calendar year
}
\description{
This function pulls and cleans data from yearly fatal traffic collision data from the Washington Traffic Safety Commission.
Data comes from the WTSC's Coded Fatal Crash data files.
}
\examples{

annual_crash_data <- process_crash_data_annual()

}
