% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatal_crash_data.R
\name{process_fars_data}
\alias{process_fars_data}
\title{National Traffic Fatalities from FARS}
\usage{
process_fars_data(safety_yrs = c(seq(2010, 2020, by = 1)))
}
\arguments{
\item{safety_yrs}{List of four digit integers for years of analysis- defaults to 2010 to 2020}
}
\value{
tibble of fatal collision metrics for the region and mpos by calendar year
}
\description{
This function pulls and cleans data from yearly fatal traffic collision data from the FARS Data.
}
\examples{
\dontrun{
annual_fars_data <- process_fars_data()}

}
