% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_data.R
\name{process_gtfs_route_data}
\alias{process_gtfs_route_data}
\title{GTFS Route Data}
\usage{
process_gtfs_route_data(year, service_change)
}
\arguments{
\item{year}{Year of gtfs data as a four digit integer - available from 2015 to current year}

\item{service_change}{Service Change period - either Fall or Spring}
}
\value{
tibble of gtfs route level data by service period year
}
\description{
This function pulls and cleans GTFS Route data from the PSRC server for use in further transit analysis.
}
\examples{
routes_2015 <- process_gtfs_route_data(year=2015, service_change = "Spring")

}
