% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_data.R
\name{process_mpo_fars_data}
\alias{process_mpo_fars_data}
\title{Summarize MPO Collision Data for RTP Dashboard}
\usage{
process_mpo_fars_data(safety_yrs = c(seq(2010, 2021, by = 1)))
}
\arguments{
\item{safety_yrs}{Vector of years in four digit integer format to analyze - defaults to 2010 to 2021}
}
\value{
tibble of annual fatal collisions and rates for Metro regions
}
\description{
This function summarizes the processed fatal collision data from FARS into a format
for the RTP Dashboard. Summaries include Total and Rate per 100k people for the 27 largest MPO's.
}
\examples{
\dontrun{
mpo_collisions <- process_mpo_fars_data(safety_yrs=c(seq(2020,2021,by=1)))}

}
