% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npmrds_data.R
\name{process_npmrds_data}
\alias{process_npmrds_data}
\title{Congested Lane Miles by Time of Day and County}
\usage{
process_npmrds_data(file_path = "X:/DSA/rtp-dashboard/NPMRDS/")
}
\arguments{
\item{file_path}{Absolute file path to pre-processed travel time files from NPMRDS}
}
\value{
tibble lane-miles of congestion by severity and time of day
}
\description{
This function cleans data from the NPMRDS output files from INRIX.
Data is pre-processed through a python sccript on data downloaded from the RTIS site
}
\examples{
\dontrun{
congested_lanes_miles <- process_npmrds_data()}

}
