% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntd_data.R
\name{process_ntd_data}
\alias{process_ntd_data}
\title{Annual Transit Metrics from the National Transit Database}
\usage{
process_ntd_data(ntd_file = NULL)
}
\arguments{
\item{ntd_file}{Sometimes FTA renames the file so you can set the name here - defaults to NULL}
}
\value{
tibble of transit related metrics for the region and mpos by calendar year
}
\description{
This function pulls and cleans data from the Monthly NTD Raw Data Release.
Data is pulled from https://www.transit.dot.gov/ntd/data-product/monthly-module-raw-data-release
}
\examples{

ntd_data <- process_ntd_data(ntd_file="September\%202022\%20Raw\%20Database_0.xlsx")

}
