% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\name{process_pums_data}
\alias{process_pums_data}
\title{Totals and Shares of Commute Travel Mode and Time from PUMS Data}
\usage{
process_pums_data(pums_yr, pums_span = 5)
}
\arguments{
\item{pums_yr}{List of PUMS data years. Each year should be the latest in non-overlapping five-year series if span is 5.}

\item{pums_span}{Integer of either 1 or 5 for the 1 or 5 year data series}
}
\value{
tibble of mean, total, and share of travel mode, time and departure time to work by race/ethnicity, county, and region
}
\description{
This function processes Public Use Microdata Sample (PUMS) estimates from the U.S. Census Bureau.
Data is pulled using PSRC's psrccensus package.
}
\examples{
\dontrun{
commute_data <- process_pums_data(pums_yr = c(2021))}

}
