% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_data.R
\name{process_safety_data}
\alias{process_safety_data}
\title{Summarized Safety data various geographies}
\usage{
process_safety_data(
  serious_injury_data_file = "X:/DSA/rtp-dashboard/serious_injury_data.csv",
  base_yr = 2010,
  wtsc_yr = 2022,
  fars_yr = 2021
)
}
\arguments{
\item{serious_injury_data_file}{Path and Name of pre-processed serious injury data file}

\item{base_yr}{Four digit integer for first year of safety data - defaults to 2010}

\item{wtsc_yr}{Four digit integer for last year of Fatal data from the WSTC - defaults to 2022}

\item{fars_yr}{Four digit integer for last year of FARS data - defaults to 2021}
}
\value{
tibble of serious and fatal collisions for use in RTP Dashboard
}
\description{
This function summarizes vehicle registration data by electrification level 
The data is pre-downloaded due to size from: https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x
}
\examples{
\dontrun{
collision_data <- process_safety_data(base_yr=2010, wtsc_yr=2022, fars_yr=2021)}

}
