% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcrtp_details.R
\docType{package}
\name{psrcrtp}
\alias{psrcrtp}
\alias{psrcrtp-package}
\title{psrcrtp: A package for data processing to feed the RTP Dashboard in the Puget Sound Region.}
\description{
Various functions to process trend data in the Central Puget Sound Region.
}
\author{
Craig Helmann \email{chelmann@psrc.org}
}
