% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{regional_housing_data}
\alias{regional_housing_data}
\title{Housing Trends from the Office of Financial Management}
\usage{
regional_housing_data(forecast_base_yr = 2018, start_year = 2010)
}
\arguments{
\item{forecast_base_yr}{Four digit integer for Base Year in Forecast Data - defaults to 2018}

\item{start_year}{Four digit integer for first year of data to use from Elmer - defaults to 2010}
}
\value{
tibble of housing trends for the region and regional geographies by calendar year
}
\description{
This function pulls and cleans data from the Intercensal Housing data from OFM.
Forecast data is from the latest VISION 2050 modeling as stored on the file server
}
\examples{
\dontrun{
hsg <- regional_housing_data()}

}
