% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_data.R
\name{summarise_collision_data}
\alias{summarise_collision_data}
\title{Summarize Collision Data for RTP Dashboard}
\usage{
summarise_collision_data(data_file)
}
\arguments{
\item{data_file}{Path and Name of pre-processed data file}
}
\value{
tibble of annual fatal and serious injury collisions and rates
}
\description{
This function summarizes the processed fatal and serious collision data into a format
for the RTP Dahsboard. Summaries include by Region, County, Race, Age Group, Gender and Mode.
}
\examples{
\dontrun{
collision_data <- summarise_collision_data(data_file="processed_collision_data.rds")}

}
