% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_registrations.R
\name{total_vehicle_registrations}
\alias{total_vehicle_registrations}
\title{All Vehicle Registrations by Electrification Level}
\usage{
total_vehicle_registrations(
  st = c("WA"),
  counties = c("King", "Kitsap", "Pierce", "Snohomish")
)
}
\arguments{
\item{st}{List of states using two digit abbreviation - defaults WA}

\item{counties}{List of County Names - defaults to PSRC counties}
}
\value{
tibble of all vehicle registrations by electrification level by month and county
}
\description{
This function pulls and cleans data from the Washington State Registration Data.
Data is pulled from https://data.wa.gov/ via the Socrata API
}
\examples{
\dontrun{
all_veh <- total_vehicle_registrations()}

}
