% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_data.R
\name{transit_stops_by_mode}
\alias{transit_stops_by_mode}
\title{GTFS Stops by Mode}
\usage{
transit_stops_by_mode(year, service_change)
}
\arguments{
\item{year}{Year of gtfs data as a four digit integer - available from 2015 to current year}

\item{service_change}{Service Change period - either Fall or Spring}
}
\value{
tibble of gtfs stop level data by service period year and mode
}
\description{
This function pulls and cleans GTFS Stop data from the PSRC server for use in further transit analysis.
}
\examples{
stops_2021 <- transit_stops_by_mode(year=2021, service_change = "Spring")

}
