% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_data.R
\name{wsdot_serious_injury_collisions}
\alias{wsdot_serious_injury_collisions}
\title{Serious Injury Collisions from WSDOT Data Files}
\usage{
wsdot_serious_injury_collisions(data_file = NULL)
}
\arguments{
\item{data_file}{Path and Name of pre-processed data file if desired - defaults to NULL}
}
\value{
tibble of annual serious injury collisions by year with city, county and tract ids
}
\description{
This function processes annual data from the summarized Injury Collision Data
from the Washington State Department of Transportation
It adds City, County and Census Tracts to the Collision Data
The final data contains collisions by gender, age, mode and roadway type
}
\examples{
\dontrun{
serious_injury_collision_data <- wsdot_serious_injury_collisions()}

}
