% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_data.R
\name{wstc_fatal_collisions}
\alias{wstc_fatal_collisions}
\title{Fatal Collisions from Washington Coded Fatal Crash (CFC) Data Files}
\usage{
wstc_fatal_collisions(data_years = seq(10, 22, by = 1))
}
\arguments{
\item{data_years}{Two digit integer (or vector of integers) for analysis years}
}
\value{
tibble of annual fatal collisions by year
}
\description{
This function processes annual data from the summarized Fatal Collision Data
from the Washington State Traffic Safety Commission
It adds City, County, Regional Geography, Centers, 2010 and 2020 Census Tracts,
Displacement Risk Score and Historically Disadvantaged Communities to the
Fatal Collision Data. The final data contains accidents by race, gender, age,
mode, roadway type, day of week, time of day, lighting and distracted driving.
}
\examples{
\dontrun{
fatals <- wstc_fatal_collisions()}

}
