% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spreadsheet.R
\name{create_public_spreadsheet}
\alias{create_public_spreadsheet}
\title{Excel Spreadsheet for Public Download of RTP Monitoring Data}
\usage{
create_public_spreadsheet(table_list)
}
\arguments{
\item{table_list}{List of data tables and spreadsheet tab names from the dashboard.}
}
\value{
spreadsheet of RTP Monitoring data
}
\description{
This function creates a Microsoft Excel spreadsheet for public download from the RTP Monitoring dashboard.
Each tab contains a single table of data from the dashboard.
}
