% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntd_data.R
\name{process_ntd_data_annual}
\alias{process_ntd_data_annual}
\title{Annual Transit Metrics from the National Transit Database}
\usage{
process_ntd_data_annual()
}
\value{
tibble of transit related metrics for the region and metro areas by calendar year
}
\description{
This function cleans data from the Monthly NTD Raw Data Release.
Data is downloaded from https://www.transit.dot.gov/ntd/data-product/monthly-module-raw-data-release
}
\examples{
\dontrun{
annual_ntd_data <- process_ntd_data_annual()}

}
