% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pums_data.R
\name{process_pums_data}
\alias{process_pums_data}
\title{Totals and Shares of Commute Travel Mode and Time from PUMS 5-Year Data}
\usage{
process_pums_data(pums_yr)
}
\arguments{
\item{pums_yr}{List of PUMS data years. Each year should be the latest in non-overlapping five-year series.}
}
\value{
tibble of mean, total, and share of travel mode and time to work by race/ethnicity, county, and region
}
\description{
This function processes 5-year Public Use Microdata Sample (PUMS) estimates from the U.S. Census Bureau.
Data is pulled using PSRC's psrccensus package.
}
\examples{
\dontrun{
commute_data <- process_pums_data(pums_yr = c(2015,2020))}

}
